+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+* [`useValidation`] Removed the use expectedFields parameter as it was never used
+
+### Fixed
+* [`useValidation`]
 
 ## [0.0.6] - 2019-08-18
 ### Changed